fix: 使用 std::sin/cos/tan 替代 sinf/cosf/tanf 解决 MSVC 兼容性

+++ Source/Runtime/Core/Private/Math.cpp
@@ -3,13 +3,14 @@
 // =============================================================================
 
 #include "Math.h"
+#include <cmath>
 
 namespace Engine
 {
 
 float Vector3::Length() const
 {
-    return sqrtf(x * x + y * y + z * z);
+    return std::sqrt(x * x + y * y + z * z);
 }
 
 Vector3 Vector3::Normalized() const
@@ -24,8 +25,8 @@ Vector3 Vector3::Normalized() const
 
 Matrix4x4 Matrix4x4::RotationX(float angle)
 {
-    float     c = cosf(angle);
-    float     s = sinf(angle);
+    float     c = std::cos(angle);
+    float     s = std::sin(angle);
     Matrix4x4 result;
     result.m[1][1] = c;
     result.m[1][2] = s;
@@ -36,8 +37,8 @@ Matrix4x4 Matrix4x4::RotationX(float angle)
 
 Matrix4x4 Matrix4x4::RotationY(float angle)
 {
-    float     c = cosf(angle);
-    float     s = sinf(angle);
+    float     c = std::cos(angle);
+    float     s = std::sin(angle);
     Matrix4x4 result;
     result.m[0][0] = c;
     result.m[0][2] = -s;
@@ -48,8 +49,8 @@ Matrix4x4 Matrix4x4::RotationY(float angle)
 
 Matrix4x4 Matrix4x4::RotationZ(float angle)
 {
-    float     c = cosf(angle);
-    float     s = sinf(angle);
+    float     c = std::cos(angle);
+    float     s = std::sin(angle);
     Matrix4x4 result;
     result.m[0][0] = c;
     result.m[0][1] = s;
@@ -60,7 +61,7 @@ Matrix4x4 Matrix4x4::RotationZ(float angle)
 
 Matrix4x4 Matrix4x4::Perspective(float fov, float aspect, float nearZ, float farZ)
 {
-    float     tanHalfFov = tanf(fov * 0.5f);
+    float     tanHalfFov = std::tan(fov * 0.5f);
     Matrix4x4 result;
     result.m[0][0] = 1.0f / (aspect * tanHalfFov);
     result.m[1][1] = 1.0f / tanHalfFov;
